refactor: restructure project into ia_sarah package
