+++ docs/tutorial.rst
@@ -4,6 +4,6 @@ Quick Start
 Install dependencies and run the application::
 
     poetry install
-    poetry run gestor-alunos
+    poetry run iasarah
 
 The interface lets you register students and training plans and export them in different formats.
